`Either` structure (part 3)
Remove the sequence under an "either" type term.

Fix parsing of types applied to multiple arguments, for example,
`(int, string) either`. During parsing, the pair `(int, string)`
becomes a "composite type" because it needs to look like an
atomic type to the parser. But it eventually becomes an
application of the "either" type to two type arguments.