Derive types for query expressions (#14)
Fixes #14

In #7 we completed type-derivation for simple expressions,
and in this change we add type-derivation for queries (`from`,
`exists`, `forall`). We handle the `compute`, `except`,
`intersect`, `union`, `group`, `into`, `through`, `unorder`
steps.

The work is not complete. Type derivation for queries that
mix ordered (`list`) and unordered (`bag`) data will require
an implementation of operator overloading, and unification
with constraints on types.

Combine the AST representation of join and scan: remove
`Join` and `JoinEq` variants and merge into the `Scan`,
`ScanEq`, `ScanExtent` variants.

Improve deduction of implicit labels in record expressions.
(Similar logic is used merging the field lists of a `group`
step with a `compute` clause.)

Disallow "on" after "join ... =" (see morel#317).

In shell and tests, print compile errors.